requirements.txt
Django==1.7
click==3.3
python-decouple==2.3
